not sure why integration test was in the frontend folder but I moved it to integration folder
